(PUP-5059) Add the --node feature to lookup CLI tool
Prior to this commit, the --node flag did not actually work as
intended, it would simply always default to the machine you were
running lookup on due to some settings not being configured properly.

Now you can specify a node with --node and the host name and lookup
will find and compile that node to obtain a scope. If you do not
specify a node then lookup will default to using the scope of the
machine you are running on (which is intended to be the master).