(maint) Don't rely on apply_manifest_on return value
Previously, these tests failed when using a rhel master and windows
agent, because apply_manifest_on will return nil when the agents array
is empty, see BKR-571.

This failure started occurring, because BKR-533 used to overwrite the
array of all hosts, making the test unsuitable for all hosts even if
there was only a single windows agent. That issue was fixed in beaker
2.25.0, and unmasked BKR-571.

However, the tests are not following best practice, which is to just
pass the block to the apply_manifest_on method, and that will yield each
result to the block.

[skip ci]