(PUP-5360) Adds support for autorelation of defined types via custom types
Custom types previously could only do an autorelation(autorequire, etc.)
of other types because that's all we looked for.  In some, possibly more
edge cases, we may need to be able to look for defined types as well.
This allows us to use autorelation if that defined type has been
included in the catalog run by using it's exact defined type name such
as :my_module::mydefine .   Previously this was only possible by stating
autorequire(:component) do ... end which is more or less the same
thing, except that this will only do the autorequire if you happen to
have that specific defined type loaded with any name in the catalog...
I think this makes more sense possibly, thoughts?
One downside to this is that the only way to do a hard autorequire is to
use a type since types aren't reliant on existing in the catalog.  So
you can currently do an autorequire(:component) { ...if
Module::DefinedType['name'] doesn't exist, fail }
Because defined types only exist if included in the catalog then this
method only allows for semi optional hard requirements such as ensuring
the name is correct, and not missing, in a defined type if another of that type exists
already.