(PUP-5089) Allow classes to be application components
This commit makes it possible for a class to be an application
component by adding the following behavior:

The parser will accept the syntax:

Class[literal] consumes Cap
Class[literal] produces Cap

The parser AST model is changed so that the LHS of a
CapabilityMapping is an Expression rather than just a String
and the validation was enhanced to cope with that.

The name of the LHS in the CapabilityMapping was changed form
'resource_name' to 'component' since that is more accurate
now.

The evaluator will do either a find_definition or a
find_hostclass depending on what type of expression that is
found in 'component'.

A host class is considered a valid producer during evaluation.