(PUP-5271) ADSI User - Do not manage unspecified password
Currently when the password policy for "Password must meet complexity
requirements" is enabled, attempts to have puppet create Windows users
without passwords fail with result `"OLE error code:800708C5 in Active
Directory. The password does not meet the password policy requirements.
Check the minimum password length, password complexity and password
history requirements."`

This behavior is expected when the password is specified but doesn't meet
the password policy, but not expected when the user has not attempted to
manage the password, e.g. leaving it empty. Correct Puppet not to try to set
the password unless it is specified.

Note that when the password policy "Minimum password length" is greater than
0, the password must always be specified. This is due to how Windows
validates new user creation and it is not possible to leave password
unspecified once that password policy is set.

Also it's important to note that when a user is specified with
`managehome => true`, the password must always be specified if it is not
an already existing user with a password.