(PUP-5342) Infer empty array as Array[Unit,0,0]
Before this, an empty array was inferred to have the type
Array[Undef, 0, 0]. This caused assigability of an empty array
to fail when assigned to an Array not accepting Undef.
By changing the type to Array[Unit, 0, 0] the empty array becomes
assignable to any Array that accpts 0 size. This because of the "magic"
trait that Unit is a cameleont type ("whatever you want it to be").

This also adds a test case for empty array assignability.