(PUP-5151) Don't allow remote file buckets to be listed
Previously the filebucket list command was implemented using the `find`
method on the `file_bucket_file` endpoint passing along `list_all` as a
query parameter. However, the filebucket is accessible to all users, so
any remote user could list the contents of the filebucket and retrieve
file content backed up by other hosts. And since retrieve and list both
use `find` where the only difference is the `find_all` query parameter,
it wasn't possible to write an ACL that allowed access to the
former, but denied access to the latter.

This commit modifies the dipper client code to raise an error when
trying to list a remote filebucket. It also modifies the
Puppet::FileBucketFile::File#list server-side code to reject remote
requests, for example, if you curl the REST API directly.