(PUP-4740) Add configured_environment to node and catalog APIs
The client includes which environment was configured on the client (if
not using the application default) as the configured_environment in node
and catalog requests. This may be used by a node classifier when
determining the environment the client should use, as added in PUP-4521.

Also adds that the transaction_uuid can be included in the node request.