(PUP-5314) Switch to allowed shell on AIX in shell test
Previously, the should_manage_shell test tried to switch the
user's shell to /bin/bash, which actually won't work on AIX
because it is not specifically allowed in /etc/security/login.cfg
and thus is not considered valid.

Modify the test to use ksh for AIX instead, which is an allowed
shell.