(PUP-5383) Don't restart puppetserver twice
Previously, we were restarting puppetserver twice during each call to
`with_puppet_running_on`, which takes about 30 seconds each time. The
first restart occurs because the test needs to change configuration,
e.g. autosign policy, destructive ssl operations, etc. The second
restart restores the original configuration and brings the server back
up. But that is useless, because we stop the server as soon as we call
`with_puppet_running_on` the next time.

This commit sets the `restart_when_done beaker` option to false globally
in AIO, which requires Beaker 2.27.0 to work correctly and is already
the minimum beaker version we depend on for acceptance.

Also two tests were calling `with_puppet_running_on` in a loop
unnecessarily, so the call was moved outside the loop.

In practice, acceptance tests on redhat-7-x86_64 dropped from 6345
seconds to 4756 seconds, saving over 25 minutes!

Finally, it doesn't make sense to backport to 3.x, since we only started
testing with puppetserver in 4.x.

[skip ci]