(PUP-5382) Validate type of nodes mapping attribute
Previously, an arbitrarily typed value could be assigned to the nodes
key of an app declaration, and the compiler would attempt to operate on
it as if it was a hash. For the sake of clarity to the user, we should
disallow the assignment of non-hash values.

This commit adds logic to verify that the value of the nodes key on an
application is either undefined, or a hash.