(maint) replace skips with expect_failure in service enable linux
The previous version of this test had some heavy-handed skip_tests which
would not remind us to fix the test when upstream bug is fixed. This
change uses expect_failure to remind us to fix the test. This came out
of some some related beaker work with this file as an example. free bird
in the hand, or something.

[skip ci]