(PUP-5380) Cache future_parser flag in environment
Prior to this commit, the Puppet.future_parser? method would load
the environment configuration from disk each time the method was
called. This had a very negative performance impact since it happens
in a lot of places.

This commit ensures that the future_parser flag is cached in the
environment instance.

All caches are cleared when the global configuration is changed.
Individual environment caches are cleared when the environment times
out.