(PUP-5294) Do not validate parameters of unevaluated consumers
When compiled with the environment compiler, consumers of capability
resources that has non-default parameters will fail with a validation
error on parameters that are expected to be consumed from the
capability. At validation time only the type and title of the
capability is known.

This commit relaxes the parameter validation when the environment
compiler is used. The compiler will detect resources that consumes
capabilities (using require or consume meta parameters) and mark such
resources as `unevaluated_consumer`. The validation of such resources
will not consider a missing but required parameter an error.