(PUP-4954) Don't initialize environment object in indirector requests unless it's used
Previously, any indirector request to a URI that had the appropriate
environment/indirector/key pathing would cause the request to
initialize an environment object. With directory environments, this
does not work - the agent probably doesn't have a dir for the
environment of the request, so initializing the environment object
fails.

This change makes initialization of the request-associated environment
object lazy, such that it only happens if the environment is actually
requested. It also makes a small change to the http indirected_routes
logic so that it does not force creation of the environment object
simply to query its name.