(PUP-5089) Add logic and test needed for class references in apps
This commit adds the logic needed to ensure that application components
in the form of class references are not evaluated unless they are
mapped to the current node in the 'nodes' attribute of a site.

Tests are also added to verify that the correct set of class references
are evaluated and that consume/produce works OK with classes in
applications.