(PUP-5060) Allow users to override facts
Add support for the --facts flag in puppet lookup. This allows the
user to give a .json or .yaml file full of fact names and values
via the --facts flag. The facts the user specified will be merged
with the existing facts for the current node. This means that if
a fact is specified in the file it's value will be overridden, but
if it is not specified, the original fact value for the node will
stay the same.