(PUP-5381) Fix access operator for Undef and Optional types
Passing a literal String as T to the Undef[T] or Optional[T] types
was handled badly in the AccessOperator and equally badly in the
unit test that verified the function (for Undef[T], there was no
unit test for Optional[T]).

This commit fixes the problem and the tests.