(PUP-5391) Fixup AIX nim provider acceptance test
This commit adds a teardown step to the nim package provider
test which removes the packages it installs once the test is
complete. This prevents failures on subsequent runs on the same
host when the test attempts to install the same package. In addition,
we now check stderr instead of stdout for the expected downgrade
error.