(PUP-5427) Don't set PATH in puppet's service scripts
Prior to this commit, some of puppet's service
scripts were setting PATH. This appears to be
unneccesary. Additionally, the specific PATH that
was being set could lead to unexpected use of the
puppet-agent's "private" utilities, such as gem.

This commit removes PATH where it was set in
puppet's service scripts.