(maint) Defer evaluation of capability mappings to after eval of applications
The fix for PUP-5275 added a second evaluation of capability mappings to ensure
evaluation after sites had been evaluated. This was still too early. The
second evaluation has to take place after applications has been evaluated
to cater for cases when the mapping isn't declared in init.pp

This commit moves the evaluation of the capability mappings to just after
the evaluation of applications.