(PUP-5395) Use lookup_options object and no global setting
This commit removes the global `hiera_advanced_parameter_bindings` and
introduces a new `lookup_options` hash.

The `lookup_options` hash replaces the previous mechanism where
the lookup_options of a key was found by doing a lookup with the key
prefixed with 'lookupoptions::'. Now, the `lookup_options` is resolved
once (per module) and merged using a hash merge. The key is then used
verbatim to find the options for a specific key in that hash.

Some refactoring was needed to utilize an adapter to cache the lookup
options. In this commit, a new LookupAdapter inherits from the
Puppet::DataProviders::DataAdapter and all lookup specific behavior
was moved to the LookupAdapter leaving the DataAdapter a pure cache
for a data hash.