(PUP-4917) Windows Packages Use QuietDisplayName
When DisplayName is not available but QuietDisplayName is, use
QuietDisplayName instead. If neither are found, continue the
behavior of returning nothing. Some software in Windows 2012
will install to QuietDisplayName instead of DisplayName, which
possibly keeps it out of the Programs and Features list, but
the value should continue to show up for Puppet.