(PUP-5387) Wait for services to finish state transitions in AIX
Previously when stopping a service, the AIX SRC provider would hit
the service with the stop signal via `stopsrc` and return immediately
as the service asynchronously shut down. In some cases, this would
cause a race condition if the service took a moment to actually stop.
In such cases, Puppet would became unaware of the actual state of the
service.

This is due to the asychronous nature of service management in AIX,
and the fact that state transitions are not atomic.

This commit adds a 60 second timeout when stopping services to check
that the service is actually stopped before returning. If a service
has not stopped by the end, it fails, since that may affect future
resources depending on the service being stopped.

For consistency, this commit also adds the same timeout for starting
services to ensure they are actually running after we start them.