(PUP-4516) Add signal name in exit message
PUP-4516 and commit c9a7e6d506768cb1c00767f785339bec512d069e introduce
a message before exiting after receiving a TERM or INT signal.

This commit adds the signal name to this error message, clarifying the
fact that we actually exit because we receive a signal and specifying
the signal.