(PUP-5425) Fix bug with phantom node in PuppetDB
When this commit is merged it will fix the issue where lookup created
a phantom node in PuppetDB. This is achieved by using the certname
of the current node rather than the string "default" to find the
current node. This removes the issue where on the second lookup run,
lookup would attempt to access the phantom node and fail because the
node does not exist.