(PUP-5441) Sanitize node's trusted info
The node passed into the compiler can be populated from
a variety of sources depending on system configuration and system
state. Previously, a node object which included a 'trusted'
parameter (as returned by, say, puppetdb or the yaml node cache)
would result in an exception during compilation due to the
need to keep $trusted immutable

To meet the need to keep top-scope $trusted immutable
while respecting the various sources of trusted data in the node
object, this commit introduces a sanitize_node method, which
deletes the 'trusted' parameter from the node object, but
resurrects it as node.trusted_data where appropriate.