(PUP-5465) Add missing support for Resource References in collection
The support for resource references in collector search expressions is
missing. This is a documented aspect of resource collectors and
previously worked in Puppet 3.

This commit adds support for resource references in collection search
expressions.

As part of the behavior, the new code converts from the POPS object into
the Puppet 3 style objects as the other internal mechanisms being
interacted with still work in with those objects.