(maint) Test using the correct catalog
Commit f066e74a changed the static compiler terminus so that it extended
the normal compiler, instead of delegating to it. In doing so the
`compiler` method was removed, causing the unit test to no longer return
a catalog of our choosing. Instead `subject.find` was always returning a
mostly empty catalog containing a main class and stages, but no file
resources. This also caused the owner, group, mode test to incorrectly
pass, because in puppet 4, numeric modes must be represented as quoted
strings.

This commit changes the test to stub the `compile` method of the normal
compiler. There is one functional change in that we always stub the
`store_content` method instead of only when `build_catalog` is called
with no arguments.

It also updates the test to expect quoted owner, group, and mode, see
PUP-4781.