(PUP-5437) Add 'lookup_options' to the lookup explainer
The explainer will currently only report the actions that lookup
performs during traversal of global data binders and providers in
order to find the data for a given key. This commit adds the lookup
of the new 'lookup_options' to the report.

Since this might make the report quite verbose, this commit also
adds an extra '--explain-data' option to the lookup CLI. This option
is mutally exclusive to '--explain' and will yield a report where the
'lookup_options' traversal is excluded.