(maint) Improve error message for xxx::data not found
Before this, the phrasing of the error message issues when an
xxx::data() function could not be found.

Now it says:

"Cannot find the function '#{name}' - required when using 'function'
data provider scheme"

The name will be either `environment::data` or `modulename::data` which
should indicate which function is not found.