(PUP-5470) Update handle_upstart test to case out Ubuntus by version
This commit modifies the confine block on the handle_upstart test to
confine out the execution of the ticket_14297_handle_upstart test based
on the platform version numbers, e.g., 15.04 or greater, as opposed to
just the platform code name, e.g., Vivid or Werewolf.  This is needed
because the platform field is populated for code names on some CI
pipelines - core Puppet - but by version number on other pipelines -
e.g., Puppet Server.