(PUP-5422) Setup signal trap before waiting for certs
When running the agent, waiting for certificates can sleep and wait for
a significant period of time. During that time a SIGTERM event can come
in (from Ctrl-C or service shutdown), causing an exception to be thrown
that prevents daemon cleanup.

Setup signal traps before waiting for certs so we ensure daemon cleanup
occurs. Update the master as well so all call paths only call
set_signal_traps once, and do so immediately after possibly daemonizing.