(maint) Use beaker's puppet helper
Previously, the test passed on Windows when using Beaker 1.x, but failed
when using 2.x, because only in 1.x did we pass the DEFAULT_GIT_ENV to
the beaker command, which ensured puppet was in the PATH.

This commit is mostly a backport of e41709e3e plus other commits made in
stable. The outcome is that we use beaker's puppet helper that knows to
wrap the puppet.bat command in cmd.exe on windows, we no longer
reference the Command::DEFAULT_GIT_ENV constant, as that doesn't exist
in beaker 2, and we single quote the less-than operator, so it's not
interpreted as redirecting stdout.

[skip ci]