(PUP-5478) Fix tests failing in Fedora 22
This commit updates two acceptance tests which were failing in Fedora
22: service_enable_linux.rb and
common_package_name_in_different_providers.rb.

The systemd provider is already the default for all versions of Fedora,
so there is no need to bump the provider.

For the common package name test, we need to use puppet-agent's private
gem binary, as rubygems is not installed on our Fedora 22 testing VMs.