(PUP-5308) Ensure future parser setting is in main section
Previously, future parser acceptance failed because `puppet config set`
does not work correctly if there is no main section (see PUP-4755),
which occurs when testing future parser with `git` style acceptance.

During the dynamic environment test, the master contained the following
puppet.conf:

    [root@u8digeb42pv9e5r ~]# cat /etc/puppet/puppet.conf
    [global]
    parser = future

    [agent]
    server = u8digeb42pv9e5r.delivery.puppetlabs.net

    [main]
    manifest = $confdir/dynamic/$environment/manifests
    modulepath = $confdir/dynamic/$environment/modules
    config_version = $confdir/static-version.sh

And future parser was not enabled:

    [root@u8digeb42pv9e5r ~]# puppet master --configprint parser
    current

This confused the dynamic environment test, which expected future parser
behavior, specifically an error message of the form:

    Could not find class ::testing_mod

This commit works around the issue by ensuring we create a main section
followed by the future parser setting. It uses beaker to introspect the
location of the puppet.conf, since it's different on windows.