(PUP-5437) Add 'lookup_options' to the lookup explainer
The explainer will currently only report the actions that lookup
performs during traversal of global data binders and providers in
order to find the data for a given key. This commit adds an option
to include the lookup of the new 'lookup_options' hash to the report.

The '--explain-options' option takes no argument can be used alone or
in combination with the '--explain' option. When used alone, and
when no keys are passed to the lookup command, the result will be
the global lookup_options hash (merge of global data binding and
environment data provider). When keys are passed, the lookup_options
will be the full hash that includes a hash found in the data provider
of the module identified by the key.