(PUP-5505) Improve performance of reading service plists
Reading text-based plist files with plutil appears to incur a lot of
overhead. Attempt to read all plists as text first, falling back to
plutil if that fails. Failing to read a binary plist appears to incur
very little overhead.

This appears to compare favorably to other approaches using
CFPropertyList, and is a much smaller change. Average times for `puppet
resource service` improved from 30 seconds to 5 seconds on a variety of
tested systems (OS X 10.9 with Ruby 1.9.3, OS X 10.10 with Ruby 2.0.0,
OS X 10.11 with Ruby 2.1.7).