(PUP-5682) Fix yumrepo param specifying false fail
Previously, there was no munge logic for YUM_BOOLEAN parameters of the
yumrepo type. Boolean false parses as a valid value for these parameters
due to the way Puppet::Parameter::Value default validation methods work.
However, "false" or Boolean false is not actually valid. Only "False".

Furthermore, recieving a Boolean false @should value will actually cause
the provider to incorrectly believe the property to be insync when it is
not.

This commit adds common munging logic to each parameter that accepts
YUM_BOOLEAN values, converting Boolean false to string false to ensure the
default validation routines work correctly and property insync is correctly
detected.

While working on this problem, another issue became apparent, which is that in
Puppet 4 "absent" passed to YUM_BOOLEAN parameters comes back as a string.
However, the provider only knows to remove the relevant line from the config
file if absent is a symbol. Therefore, the munge method also converts string
absent to symbol absent.