(PUP-5511) Make lookup command accept all valid choices for --merge
The lookup command didn't accept the merge strategy 'first'. Probably
because this strategy was the default but could not be explicitly
specified in earlier versions.
To fix this, and also cater for other future additions, this commit
changes the validation of the --merge parameter so that it uses the
MergeStrategies::strategy_keys method instead of a hard-coded set.