(PUP-5285) Update systemd provider to detect all unit types
Prior to this commit, when searching for discoverable systemd
services we filtered our results down to only units of the 'service'
type. There are several other valid types, such as timer and socket
which can be managed as well. This commit updates our systemctl
invocation to allow any unit type to be returned in its output.

Note that managing systemd unit types other than the service type
requires the fully-qualified service name to be provided. For example,
to manage the unit 'foo.timer', Puppet must be provided with 'foo.timer'
rather than just 'foo'. Otherwise, the provider will default to managing
the service unit type.