(PUP-5060) Fix lookup --facts bug
Prior to this commit, the facts passed in via the --facts were not
being promoted to topscope because they were not being added to the
node parameters. This meant they would not be found in a lookup.

Now, facts in the given fact file are merged into the node parameters
so that when the node is compiled they will be promoted to topscope.