(PUP-5548) Ensure systemctl is present before using it for service status
Currently, the Debian service provider uses systemctl to determine
service status on Debian 8 and Ubuntu 15.04, which use systemd by
default. This is necessary due to the way masked services behave
when queried with the sysvint-systemd compatibility layer.

It is possible to remove systemd from Debian 8 and Ubuntu 15.04
and only use init. In such a case the Debian provider will error
when trying to use systemctl. This commit updates the provider to
check that systemctl exists on the system before attempting to use
it to determine service status. If it is not present, the provider
falls back to the service init script.