(PUP-5548) Use `service` in debian provider to interact with all init systems
In the Debian family, the `service` command provides an
abstraction around available init systems, allowing it to be used
to properly determine service state regardless of which init system
is active on the system.

This commit updates the Debian provider's `statuscmd` method to use
`service` at all times, allowing the provider to properly query
services regardless of OS version or init system in place.