(PUP-5554) Improve parser performance on JRuby
The lex_token method of the Puppet::Pops::Parser::Lexer2 class is
heavily optimized to run fast on MRI. It's been discovered that this
optimization has a fairly negative impact when running with JRuby.

This commit is a rewrite of the lex_token method with emphasis on
creating smaller blocks of code (hash with lambda approach).

The following can be measured when parsing a site containing
1000 nodes with 50 common modules from the forge repeatedly (50
times).

36% less CPU time with JRuby (tested using JRuby 9.0.4)
8% less CPU time with MRI (tested on 1.9.3, 2.1.1, and 2.2.1).

MRI is still 23% faster than JRuby but it used to be 48% faster.