(PUP-5554) Improve parser factory performance
This commit optimizes the parser factory. Profile data from runs
with JRuby was used to pinpoint some bottlenecks so the performance
increase is more apparent with JRuby than with MRI. But both will
benefit. Using the same way of measuring as in the prior commit
we now get:

55% less CPU time with JRuby
13% less CPU time with MRI

The MRI parser is now only 11% faster than JRuby.