(maint) PkgNG: Cache version results
Without this change, 'pkg version' is called once for each package found
on the system to determine if it is the latest version.  As a result of
using the '-R' flag, this reaches out to the repositories for this
information once for each package, which can contribute considerably to
the total run of the agent.

Here we cache the results of the initial query in a class variable to
ensure that only one request is made to retrieve the full list of
version information for all discovered packages.