(PUP-5509) Ensure error context for non-numeric index error
No error context (file, line, position) was provided for the case
when a non numeric access operator was used on a string. It was
however provided when the same operator was used on an array. This
was caused by an incorrect passing of the semantic parameter.
This commit fixes that.