(PUP-1376) Allow tag negation
This commit is a first attempt at achieving 'not tags' or tag negation.
I've called them 'skip tags' in this PR. This adds the ability to
specify a list of tags to be excluded from the catalog. I'm not entirely
sure this is complete, so I'd like some feedback if this is a desirable
feature (I think it is), and if this implementation is feasible.