(PUP-5608) Remove unneeded stubs which borked ruby 2.3
Previously this test was stubbing File.open and File.chown. The
latter was because of the former, and the former was put in before
we used a proper Tempfile-based approach for generating temp files.
Now that we do, preventing the creation of temp files isn't a
concern.

In ruby 2.3, Tempfile returns nil if File.open is stubbed, which
caused this test to fail with the stubs in place. I didn't track
down the change to Tempfile, but it doesn't seem unreasonable that
it would rely on File.open.