(PUP-5608) Prevent SemVer <=> method from throwing
This commit is part of adding ruby 2.3.0 support to puppet.

Note that prior to ruby 2.3.0, if a <=> method threw an exception, ruby
would silently rescue the exception and return nil from <=> (which causes
the derived == comparison to return false). Starting in ruby 2.3.0, this
behavior changed and the exception is actually thrown. Some comments at:
https://bugs.ruby-lang.org/issues/7688

This change ensures that the <=> won't throw an exception by avoiding
the constructor for invalid version parameters (because the constructor
is the one place that will throw an exception).