(PUP-5617) enable keepalive on the TCP socket opened with the master
This will allow the socket with the master to get closed when the
network connection breaks unexpectedly, and the agent to terminate with
an error, rather than waiting forever.

According to linux's tcp(7) manpage, such zombies connections get closed
by default after a bit more than 2 hours of idleness. Configuring this
is possible but not considered portable.