(maint) PkgNG: avoid provider caching
This work shifts the method calls to avoid caching the results for the
current package version information in the provider.  Now we retrieve
the package version list from the #instances method, and pass that into
the get_latest_version() method to filter down the results.

Now get_version_list should only be called once and caching is not
necessary.