(PUP-5583) Allow host configs to be generated at runtime
This commit adds a dependency on beaker-hostgenerator in the
acceptance Gemfile, which may be overridden using the
`BEAKER_HOSTGENERATOR_VERSION` environment variable.

The acceptance Rakefile will continue to use the `CONFIG` environment
variable, which must point to an existing host config file.

If `TEST_TARGET` is specified, it will override `CONFIG`, and the Rakefile
will use beaker-hostgenerator to dynamically generate a host config. The
host config is generated in the `acceptance/tmp` directory, as
opposed to `/tmp` so that it doesn't conflict with concurrent acceptance
tests running on the same CI coordinator. It also appends SecureRandom
to the hostfile, because it seems like a good idea. The resulting host
config is copied to the `log/latest/config.yml`.

The Rakefile accepts an optional `MASTER_TEST_TARGET` to override the
default master platform. By default, it will use redhat7-64m.

For example, to run against a tagged build:

    $ bundle exec rake ci:test:aio TEST_TARGET=redhat7-64a SHA=1.3.2

To run against a non-tagged commit:

    $ bundle exec rake ci:test:aio TEST_TARGET=redhat7-64a \
      SHA=<full sha> SUITE_VERSION=<git describe>

To run against an alternate master platform:

    $ bundle exec rake ci:test:aio MASTER_TEST_TARGET=debian8-64m \
      TEST_TARGET=redhat7-64a SHA=1.3.2

This commit doesn't remove the static host configs as CI jobs need to be
updated first.