(maint) Bump travis matrix to ruby 2.3.0
In the run up to the 2.3.0 release, .travis.yml was
configured to test against ruby-head. Now that 2.3.0 has
been released, this commit just replaces ruby-head testing
with 2.3.0 testing. (The commits in support of ruby 2.3 went
in earlier.)

Note that we won't continue to test against ruby-head for now.
Instead, next year around this time, we can enable ruby-head
again in the run up to next year's Dec 25 ruby release.